+++ src/frontend/apps/impress/src/features/docs/doc-versioning/components/ModalConfirmationVersion.tsx
@@ -1,11 +1,13 @@
 import {
   Button,
+  ButtonElement,
   Modal,
   ModalSize,
   VariantType,
   useToastProvider,
 } from '@gouvfr-lasuite/cunningham-react';
 import { useRouter } from 'next/router';
+import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { Box, Text } from '@/components';
@@ -15,6 +17,7 @@ import {
   useProviderStore,
   useUpdateDoc,
 } from '@/docs/doc-management/';
+import { useFocusOnMount } from '@/hooks';
 
 import { useDocVersion } from '../api';
 import { KEY_LIST_DOC_VERSIONS } from '../api/useDocVersions';
@@ -33,6 +36,7 @@ export const ModalConfirmationVersion = ({
   docId,
   versionId,
 }: ModalConfirmationVersionProps) => {
+  const cancelButtonRef = useRef<ButtonElement>(null);
   const { data: version } = useDocVersion({
     docId,
     versionId,
@@ -64,6 +68,8 @@ export const ModalConfirmationVersion = ({
     },
   });
 
+  useFocusOnMount(cancelButtonRef);
+
   return (
     <Modal
       isOpen
@@ -73,6 +79,7 @@ export const ModalConfirmationVersion = ({
       rightActions={
         <>
           <Button
+            ref={cancelButtonRef}
             aria-label={`${t('Cancel')} - ${t('Warning')}`}
             variant="secondary"
             fullWidth