+++ src/frontend/apps/impress/src/hooks/index.ts
@@ -1,4 +1,6 @@
 export * from './useClipboard';
 export * from './useCmdK';
 export * from './useDate';
+export * from './useFocusOnMount';
 export * from './useKeyboardAction';
+export * from './useRestoreFocus';